+++ amplify/backend/function/expressServer/src/routes/portfolioRoutes.js
@@ -1,0 +1,8 @@
+const express = require("express")
+const router = express.Router()
+const { createPortfolio } = require('../controllers/portfolioController')
+const auth = require('../middleware/authMiddleware');
+
+router.post('/createPortfolio', auth, createPortfolio)
+
+module.exports = router
